Move Encryption and Transient to helpers directory
